pyproject.toml
[project]
name = "medical-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"langchain==0.3.26",
	"sentence-transformers==4.1.0",
	"pypdf==5.6.1",
	"python-dotenv==1.1.0",
	"langchain-pinecone==0.2.8",
	"flask==3.1.1",
	"langchain-openai==0.3.24",
	"langchain-community==0.3.26"
	]
